Made constraint file share scope
