+++ src/stores/mutations.js
@@ -1,0 +1,3 @@
+import { mutations } from './constraints';
+
+export const APPLY_CONSTRAINT = mutations.APPLY_CONSTRAINT;
