+++ src/stores/store.js
@@ -1,8 +1,8 @@
 import Vuex from 'vuex';
 import * as actions from './actions';
+import * as mutations from './mutations';
 import * as getters from './getters';
 import {createProductStore} from './modules/products';
-import {createConstraintStore} from './modules/constraints';
 import {createCart} from './modules/cart';
 import { createLogger } from '../services/logger';
 
@@ -10,16 +10,15 @@ const debug = process.env.NODE_ENV !== 'production';
 
 export function createStore (config, shopifyClient) {
   const products = createProductStore(config, shopifyClient);
-  const constraints = createConstraintStore(config);
   const cart = createCart(shopifyClient);
 
   return new Vuex.Store({
     actions,
+    mutations,
     getters,
     modules: {
       cart,
-      products,
-      constraints
+      products
     },
     strict: debug,
     plugins: debug ? [createLogger()] : []
